simplify runner lease sidecar store structure
